backend/requirements.txt
asgiref==3.7.2
Django==4.2
django-filter==23.3
djangorestframework==3.14.0
factory-boy==3.3.0
Faker==19.6.2
gunicorn==21.2.0
mypy==1.5.1
mypy-extensions==1.0.0
packaging==23.1
psycopg2==2.9.8
psycopg2-binary==2.9.8
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.8.0
whitenoise==6.5.0
